Web copyable modal (#37)
* Changed log level to o warnings only

* Add modal window that displays log

I wasn't sure how it would impact HSC as a whole to add a new menu
option to summon the modal so I added a button in the top left that's
only visible on hover. It's a hacky design and there's probably a
better way.

* Fixed comment

* Make the JS functions compatible with the future Rust menu

Removed the hidden button in the upper right and replaced it with
functions that will be called from the Rust end. `loadPuzzleLog` is
not yet defined.

* Use promises

Copying and pasting has been generalized to all text and pasting now
returns a promise.

* Add fallback copy/paste

* Update changelog

* Add newline back to .cargo/config.toml

---------

Co-authored-by: HactarCE <6060305+HactarCE@users.noreply.github.com>